Add test script
This script is not intended to do a very thorough test of
luaphonenumber; it's more like a 'smoke test' which can be used
after building a binary. It is based on the code samples in README.md.

Speaking of which, running the test script revealed that the output
for one of the code samples in README.md was not the same as shown.